% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{date_format_jy}
\alias{date_format_jy}
\title{A date format helper to show the year in January}
\usage{
date_format_jy(x, date_label = "\%b")
}
\arguments{
\item{x}{vector of date class}

\item{date_label}{the formar of date}
}
\description{
A date format helper to show the year in January
}
