% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{date_to_ym}
\alias{date_to_ym}
\title{Date to period string}
\usage{
date_to_ym(d = as.Date(c("2009-02-01", "2019-12-01")))
}
\arguments{
\item{d}{yyyy-mm-dd or any valid date or string coercible to date}
}
\description{
Date to period string
}
\examples{
dtp <- date_to_ym(as.Date(c("2009-02-01", "2019-12-01")))
dtp
class(dtp)
}
