% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot2.R
\name{scale_x_date_year}
\alias{scale_x_date_year}
\alias{scale_x_date_yyyymm}
\title{Auxiliar shortcut scales with opinionated pre-sets}
\usage{
scale_x_date_year(date_labels = "\%Y", ...)

scale_x_date_yyyymm(date_labels = "\%Y\%m", ...)
}
\arguments{
\item{date_labels}{Date labels}

\item{...}{More arguments for `scale_x_date`}
}
\description{
Auxiliar shortcut scales with opinionated pre-sets
}
