% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym_add_months}
\alias{ym_add_months}
\title{Add months to a year month}
\usage{
ym_add_months(ym = c(201201, 201303), months = c(1, -12))
}
\arguments{
\item{ym}{ym}

\item{months}{months}
}
\description{
Add months to a year month
}
\examples{
ym_add_months(c(200902, 201912), months = c(1, -12))
}
