% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym_diff}
\alias{ym_diff}
\title{Year month string to date}
\usage{
ym_diff(ym = c(200902, 201912), ym2 = c(200901, 201712))
}
\arguments{
\item{ym}{ym}

\item{ym2}{ym2}
}
\description{
Year month string to date
}
\examples{

ym_diff(c(200902, 201912), c(200901, 201712))

}
\references{
http://stackoverflow.com/questions/1995933/number-of-months-between-two-dates/26640698#26640698
}
