% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym_div}
\alias{ym_div}
\title{Year month to semester, quarter, trimester}
\usage{
ym_div(ym = format(ymd(20170101) + months(0:11), "\%Y\%m"), ng = 4)
}
\arguments{
\item{ym}{ym}

\item{ng}{number of groups

ym_seq(201303, 201512)

seqym <- ym_seq(201303, 201512)
seqym

ym_div(seqym, ng = 3)
ym_div(seqym, ng = 4)}
}
\description{
Year month to semester, quarter, trimester
}
