% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym_format}
\alias{ym_format}
\title{Format a year month into a specific date format
A wrapper for ym_to_date and format:}
\usage{
ym_format(ym = c(200902, 201912), format = "\%Y\%m\%d", day = 1)
}
\arguments{
\item{ym}{ym}

\item{format}{format}

\item{day}{day}
}
\description{
Format a year month into a specific date format
A wrapper for ym_to_date and format:
}
\examples{

ym_format(c(200902, 201912), "\%B, \%Y")

}
