% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym_seq}
\alias{ym_seq}
\title{Sequence from}
\usage{
ym_seq(from = 200903, to = 201512, by = "month")
}
\arguments{
\item{from}{ym}

\item{to}{ym}

\item{by}{Same argument in `seq.Date`, can be "month" or "year".}
}
\description{
Sequence from
}
\examples{

ym_seq(201303, 201512)

}
