% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ym.R
\name{ym_to_date}
\alias{ym_to_date}
\title{Year month string to date}
\usage{
ym_to_date(ym = c(200902, 201912), day = 1)
}
\arguments{
\item{ym}{ym}

\item{day}{day}
}
\description{
Year month string to date
}
\examples{

dts <- ym_to_date(c(200902, 201912))
dts
class(dts)

}
