% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yyyymm-package.R
\name{yyyymm-exports}
\alias{yyyymm-exports}
\title{yyyymm exported operators and S3 methods
The following functions are imported and then re-exported
from the gravity package to avoid listing Depends of yyyymm.}
\description{
yyyymm exported operators and S3 methods
The following functions are imported and then re-exported
from the gravity package to avoid listing Depends of yyyymm.
}
\keyword{internal}
